\name{phenotypinglmem-package}
\alias{phenotypinglmem-package}
\alias{phenotypinglmem}
\docType{package}
\title{
\packageTitle{phenotypinglmem}
}
\description{
\packageDescription{phenotypinglmem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{phenotypinglmem}
\packageIndices{phenotypinglmem}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{phenotypinglmem}

Maintainer: \packageMaintainer{phenotypinglmem}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
